requirements.txt
absl-py==0.15.0
ale-py==0.7.1
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2025.1.31
chainer==7.8.1
charset-normalizer==2.0.12
clang==5.0
cloudpickle==2.2.1
cntk==2.7
cntk-gpu==2.7
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
filelock==3.4.1
flatbuffers==1.12
future==1.0.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
gym==0.21.0
h5py==3.1.0
idna==3.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.1
lxml==5.3.1
malmo==0.36.0.0
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.0.7
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.5.4
seaborn==0.11.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
torch==1.10.2
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.26.20
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.12.1
zipp==3.6.0